move save practice button to components
